+++ RELEASENOTES.md
@@ -62,6 +62,8 @@
         ([#3750](https://github.com/google/ExoPlayer/issues/3750)).
     *   Add a way to override ad media MIME types
         ([#7961)(https://github.com/google/ExoPlayer/issues/7961)).
+    *   Fix truncating large cue points in microseconds
+        ([#8067](https://github.com/google/ExoPlayer/issues/8067)).
 
 ### 2.12.0 (2020-09-11) ###